MS-66: Refactor getAvailableDates logic
