+++ controllers/candidate/sessions.js
@@ -1,7 +1,5 @@
 const Sessions = require('../../models/Sessions');
-const Psychologists = require('../../models/Psychologists');
-const { checkEmptyTimeRange, getAvailability, getCurrentWeek } = require('../../helpers');
-const { weekDays } = require('../../helpers');
+const { getCurrentWeek, getAvailableDates } = require('../../helpers');
 
 const notFoundTxt = 'Session not found with ID:';
 
@@ -56,39 +54,10 @@ const remove = (req, res) => {
     .catch((err) => res.status(500).json({ msg: `Error: ${err}`, error: true }));
 };
 
-const getAvailableDates = async (req, res) => {
-  const availableDates = [];
+const getAvailableSessions = async (req, res) => {
+  const availableDates = await getAvailableDates();
   const currentWeek = getCurrentWeek();
-  const psychologists = await Psychologists.find();
-  await Promise.all(
-    psychologists.map(async (psychologist) => {
-      if (checkEmptyTimeRange(psychologist.timeRange)) {
-        const availableTimeRange = getAvailability(psychologist.timeRange);
-        const availability = [];
-        const sessions = await Sessions.find({
-          idPsychologist: psychologist._id,
-          status: 'pending',
-        });
-        currentWeek.forEach((day) => {
-          const existingSessions = sessions.filter(
-            (session) => weekDays[session.date.getDay()] === day.day,
-          );
-          let availableHours = availableTimeRange[day.day];
-          existingSessions.forEach((session) => {
-            availableHours = availableHours.filter((hour) => hour !== session.time);
-          });
-          availability.push({ day: day.day, number: day.number, hours: availableHours });
-        });
-        const element = {
-          id: psychologist._id,
-          name: `${psychologist.firstName} ${psychologist.lastName}`,
-          availability,
-        };
-        availableDates.push(element);
-      }
-    }),
-  );
-  res.status(200).json({ availableDates, currentWeek });
+  return res.status(200).json({ availableDates, currentWeek });
 };
 
 module.exports = {
@@ -96,5 +65,5 @@ module.exports = {
   getById,
   edit,
   remove,
-  getAvailableDates,
+  getAvailableSessions,
 };
